requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
langchain==0.1.1
langchain-community==0.0.10
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pypdf==3.17.1
python-dotenv==1.0.0
pandas==2.0.3
plotly==5.17.0
reportlab==4.0.9